:recycle: refactor probe tests
